IVYPORTAL-3932 - Various or urgent requests in S140: Express starts in global search

+++ AxonIvyPortal/PortalConnector/src/ch/ivy/ws/addon/service/ProcessStartServiceImpl.java
@@ -12,6 +12,7 @@ import ch.ivy.ws.addon.WSException;
 import ch.ivy.ws.addon.bo.ProcessStartServiceResult;
 import ch.ivy.ws.addon.transformer.IvyProcessStartTransformer;
 import ch.ivy.ws.addon.types.IvyProcessStart;
+import ch.ivyteam.ivy.Advisor;
 import ch.ivyteam.ivy.application.IApplication;
 import ch.ivyteam.ivy.environment.Ivy;
 import ch.ivyteam.ivy.security.ISecurityContext;
@@ -22,6 +23,7 @@ import ch.ivyteam.ivy.workflow.IWorkflowSession;
 /**
  * Default implementation for the process start service
  */
+@SuppressWarnings("restriction")
 public class ProcessStartServiceImpl extends AbstractService implements IProcessStartService {
 
   @Override
@@ -53,7 +55,7 @@ public class ProcessStartServiceImpl extends AbstractService implements IProcess
                   }
                 }
               } finally {
-                if (workflowSession != null && application != null) {
+                if (workflowSession != null && application != null && Advisor.getAdvisor().isServer()) {
                   ISecurityContext securityContext = application.getSecurityContext();
                   securityContext.destroySession(workflowSession.getIdentifier());
                 }
@@ -85,11 +87,15 @@ public class ProcessStartServiceImpl extends AbstractService implements IProcess
 
   private IWorkflowSession getWorkflowSession(ProcessSearchCriteria searchCriteria, IApplication application)
       throws Exception {
+    if (searchCriteria.hasInvolvedUsername() && Advisor.getAdvisor().isServer()) {
+      IWorkflowSession givenUserWorkflowSession =
+          findUserWorkflowSession(searchCriteria.getInvolvedUsername(), application);
+      return givenUserWorkflowSession;
+    }
+    
     IWorkflowSession systemUserWorkflowSession =
         Ivy.wf().getWorkflowSession(application.getSecurityContext().getSystemUserSession());
-    IWorkflowSession givenUserWorkflowSession =
-        findUserWorkflowSession(searchCriteria.getInvolvedUsername(), application);
-    return searchCriteria.hasInvolvedUsername() ? givenUserWorkflowSession : systemUserWorkflowSession;
+    return systemUserWorkflowSession;
   }
 
   private static List<IApplication> getAllApplications() {
